+++ src/aac_datasets/__init__.py
@@ -2,7 +2,7 @@
 # -*- coding: utf-8 -*-
 
 """
-Automated Audio Captioning Datasets package.
+Audio Captioning Datasets package.
 """
 
 from .datasets.audiocaps import AudioCaps
@@ -14,4 +14,4 @@ __author__ = "Etienne Labbé (Labbeti)"
 __maintainer__ = "Etienne Labbé (Labbeti)"
 __name__ = "aac-datasets"
 __status__ = "Development"
-__version__ = "0.2.0"
+__version__ = "0.3.0"